type replacement

+++ src/common/CPtrList.cpp
@@ -6,7 +6,7 @@ void CVoidPtrList::AddFrontItem(void *inPtr)
 		return;
 
 	void *prevHead = mHead;
-	*(void **)((uint)inPtr + mOffset) = prevHead;
+	*(void **)((size_t)inPtr + mOffset) = prevHead;
 	mHead = inPtr;
 
 	if (!prevHead)
@@ -20,12 +20,12 @@ void CVoidPtrList::AddBackItem(void *inPtr)
 	if (!inPtr)
 		return;
 
-	*(void **)((uint)inPtr + mOffset) = NULL;
+	*(void **)((size_t)inPtr + mOffset) = NULL;
 
 	if (!mTail)
 		mHead = inPtr;
 	else
-		*(void **)((uint)mTail + mOffset) = inPtr;
+		*(void **)((size_t)mTail + mOffset) = inPtr;
 
 	mTail = inPtr;
 	mCount++;
